backend/requirements.txt
fastapi==0.95.2
uvicorn==0.27.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==1.10.13
sqlalchemy==1.4.51
langchain==0.0.350
langchain-community==0.0.8
langchain-core==0.1.8
chromadb==0.4.15
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pdfminer.six==20231228
ollama==0.1.0
pydantic-core==2.14.6
typing-extensions==4.9.0